+++ lib/bob/api/employee/documents.rb
@@ -10,6 +10,10 @@ module Bob
       def self.add_private_document(employee_id, params = {})
         post("docs/people/#{employee_id}/confidential", params, use_api_key: true)
       end
+
+      def self.upload_public_document(employee_id, file_path)
+        post_file("docs/people/#{employee_id}/shared/upload", file_path)
+      end
     end
   end
 end